Started Astar
Started and laid all the groundwork for the Astar shortest path search.
Update path to have a parent node in order to backtrack during the
search. Will finish this as soon as my thesis report gets submitted.

Also starting to see the need for a continuitty path checking function,
some background can be found at
http://en.wikipedia.org/wiki/Connected_component_(graph_theory)

If you have a second you might look into it. Since 'connected' is the
check we are doing most often it might be important.